Fix WTF. (Java's error reporting leaves something to be desired.)
