+++ server-webapp/src/main/java/org/taverna/server/master/localworker/LocalWorkerState.java
@@ -25,15 +25,11 @@ import java.io.FilenameFilter;
 
 import javax.annotation.PostConstruct;
 import javax.jdo.annotations.PersistenceAware;
-import javax.jdo.annotations.PersistenceCapable;
-import javax.jdo.annotations.Persistent;
-import javax.jdo.annotations.PrimaryKey;
 
 import org.springframework.beans.factory.annotation.Required;
 import org.taverna.server.master.common.Status;
 import org.taverna.server.master.defaults.Default;
 import org.taverna.server.master.utils.JDOSupport;
-import org.taverna.server.master.worker.PolicyLimits;
 
 /**
  * The persistent state of a local worker factory.
@@ -42,7 +38,7 @@ import org.taverna.server.master.worker.PolicyLimits;
  */
 @PersistenceAware
 public class LocalWorkerState extends JDOSupport<PersistedState> implements
-		LocalWorkerModel, PolicyLimits {
+		LocalWorkerModel {
 	public LocalWorkerState() {
 		super(PersistedState.class);
 	}
@@ -388,353 +384,4 @@ public class LocalWorkerState extends JDOSupport<PersistedState> implements
 
 		loadedState = true;
 	}
-}
-
-// WARNING! If you change the name of this class, update persistence.xml as
-// well!
-@PersistenceCapable(table = "LOCALWORKERSTATE__PERSISTEDSTATE")
-class PersistedState implements LocalWorkerModel, PolicyLimits {
-	static PersistedState makeInstance() {
-		PersistedState o = new PersistedState();
-		o.ID = KEY;
-		return o;
-	}
-
-	@PrimaryKey(column = "ID")
-	protected int ID;
-
-	static final int KEY = 32;
-
-	@Persistent
-	private int defaultLifetime;
-	@Persistent
-	private int maxRuns;
-	@Persistent
-	private String factoryProcessNamePrefix;
-	@Persistent
-	private String executeWorkflowScript;
-	@Persistent(serialized = "true")
-	private String[] extraArgs;
-	@Persistent
-	private int waitSeconds;
-	@Persistent
-	private int sleepMS;
-	@Persistent
-	private String serverWorkerJar;
-	@Persistent
-	private String serverForkerJar;
-	@Persistent
-	private String passwordFile;
-	@Persistent
-	private String javaBinary;
-	@Persistent
-	private int registryPort;
-	@Persistent
-	private String registryHost;
-	@Persistent
-	private int operatingLimit;
-
-	@Override
-	public void setDefaultLifetime(int defaultLifetime) {
-		this.defaultLifetime = defaultLifetime;
-	}
-
-	@Override
-	public int getDefaultLifetime() {
-		return defaultLifetime;
-	}
-
-	@Override
-	public void setMaxRuns(int maxRuns) {
-		this.maxRuns = maxRuns;
-	}
-
-	@Override
-	public int getMaxRuns() {
-		return maxRuns;
-	}
-
-	@Override
-	public void setFactoryProcessNamePrefix(String factoryProcessNamePrefix) {
-		this.factoryProcessNamePrefix = factoryProcessNamePrefix;
-	}
-
-	@Override
-	public String getFactoryProcessNamePrefix() {
-		return factoryProcessNamePrefix;
-	}
-
-	@Override
-	public void setExecuteWorkflowScript(String executeWorkflowScript) {
-		this.executeWorkflowScript = executeWorkflowScript;
-	}
-
-	@Override
-	public String getExecuteWorkflowScript() {
-		return executeWorkflowScript;
-	}
-
-	@Override
-	public void setExtraArgs(String[] extraArgs) {
-		this.extraArgs = extraArgs;
-	}
-
-	@Override
-	public String[] getExtraArgs() {
-		return extraArgs;
-	}
-
-	@Override
-	public void setWaitSeconds(int waitSeconds) {
-		this.waitSeconds = waitSeconds;
-	}
-
-	@Override
-	public int getWaitSeconds() {
-		return waitSeconds;
-	}
-
-	@Override
-	public void setSleepMS(int sleepMS) {
-		this.sleepMS = sleepMS;
-	}
-
-	@Override
-	public int getSleepMS() {
-		return sleepMS;
-	}
-
-	@Override
-	public void setServerWorkerJar(String serverWorkerJar) {
-		this.serverWorkerJar = serverWorkerJar;
-	}
-
-	@Override
-	public String getServerWorkerJar() {
-		return serverWorkerJar;
-	}
-
-	@Override
-	public void setJavaBinary(String javaBinary) {
-		this.javaBinary = javaBinary;
-	}
-
-	@Override
-	public String getJavaBinary() {
-		return javaBinary;
-	}
-
-	@Override
-	public void setRegistryPort(int registryPort) {
-		this.registryPort = registryPort;
-	}
-
-	@Override
-	public int getRegistryPort() {
-		return registryPort;
-	}
-
-	@Override
-	public void setRegistryHost(String registryHost) {
-		this.registryHost = registryHost;
-	}
-
-	@Override
-	public String getRegistryHost() {
-		return registryHost;
-	}
-
-	@Override
-	public void setServerForkerJar(String serverForkerJar) {
-		this.serverForkerJar = serverForkerJar;
-	}
-
-	@Override
-	public String getServerForkerJar() {
-		return serverForkerJar;
-	}
-
-	@Override
-	public void setPasswordFile(String passwordFile) {
-		this.passwordFile = passwordFile;
-	}
-
-	@Override
-	public String getPasswordFile() {
-		return passwordFile;
-	}
-
-	@Override
-	public void setOperatingLimit(int operatingLimit) {
-		this.operatingLimit = operatingLimit;
-	}
-
-	@Override
-	public int getOperatingLimit() {
-		return operatingLimit;
-	}
-}
-
-/**
- * Profile of the getters and setters in a worker system. Ensures that the
- * persisted state matches the public view on the state model at least fairly
- * closely.
- * 
- * @author Donal Fellows
- */
-interface LocalWorkerModel extends PolicyLimits {
-
-	/**
-	 * @param defaultLifetime
-	 *            how long a workflow run should live by default, in minutes.
-	 */
-	public abstract void setDefaultLifetime(int defaultLifetime);
-
-	/**
-	 * @return how long a workflow run should live by default, in minutes.
-	 */
-	public abstract int getDefaultLifetime();
-
-	/**
-	 * @param maxRuns
-	 *            the maximum number of extant workflow runs
-	 */
-	public abstract void setMaxRuns(int maxRuns);
-
-	/**
-	 * @param factoryProcessNamePrefix
-	 *            the prefix used for factory processes in RMI
-	 */
-	public abstract void setFactoryProcessNamePrefix(
-			String factoryProcessNamePrefix);
-
-	/**
-	 * @return the prefix used for factory processes in RMI
-	 */
-	public abstract String getFactoryProcessNamePrefix();
-
-	/**
-	 * @param executeWorkflowScript
-	 *            the script to run to actually run a workflow
-	 */
-	public abstract void setExecuteWorkflowScript(String executeWorkflowScript);
-
-	/**
-	 * @return the script to run to actually run a workflow
-	 */
-	public abstract String getExecuteWorkflowScript();
-
-	/**
-	 * @param extraArgs
-	 *            the extra arguments to pass into the workflow runner
-	 */
-	public abstract void setExtraArgs(String[] extraArgs);
-
-	/**
-	 * @return the extra arguments to pass into the workflow runner
-	 */
-	public abstract String[] getExtraArgs();
-
-	/**
-	 * @param waitSeconds
-	 *            the number of seconds to wait for subprocesses to start
-	 */
-	public abstract void setWaitSeconds(int waitSeconds);
-
-	/**
-	 * @return the number of seconds to wait for subprocesses to start
-	 */
-	public abstract int getWaitSeconds();
-
-	/**
-	 * @param sleepMS
-	 *            milliseconds to wait between polling for a started
-	 *            subprocess's status
-	 */
-	public abstract void setSleepMS(int sleepMS);
-
-	/**
-	 * @return milliseconds to wait between polling for a started subprocess's
-	 *         status
-	 */
-	public abstract int getSleepMS();
-
-	/**
-	 * @param serverWorkerJar
-	 *            the full path name of the file system access worker
-	 *            subprocess's implementation JAR
-	 */
-	public abstract void setServerWorkerJar(String serverWorkerJar);
-
-	/**
-	 * @return the full path name of the file system access worker subprocess's
-	 *         implementation JAR
-	 */
-	public abstract String getServerWorkerJar();
-
-	/**
-	 * @param javaBinary
-	 *            the full path name to the Java binary to use
-	 */
-	public abstract void setJavaBinary(String javaBinary);
-
-	/**
-	 * @return the full path name to the Java binary to use
-	 */
-	public abstract String getJavaBinary();
-
-	/**
-	 * @param registryPort
-	 *            what port is the RMI registry on
-	 */
-	public abstract void setRegistryPort(int registryPort);
-
-	/**
-	 * @return what port is the RMI registry on
-	 */
-	public abstract int getRegistryPort();
-
-	/**
-	 * @param registryHost
-	 *            what host (network interface) is the RMI registry on
-	 */
-	public abstract void setRegistryHost(String registryHost);
-
-	/**
-	 * @return what host (network interface) is the RMI registry on
-	 */
-	public abstract String getRegistryHost();
-
-	/**
-	 * @param serverForkerJar
-	 *            the full path name of the impersonation engine's
-	 *            implementation JAR
-	 */
-	public abstract void setServerForkerJar(String serverForkerJar);
-
-	/**
-	 * @return the full path name of the impersonation engine's implementation
-	 *         JAR
-	 */
-	public abstract String getServerForkerJar();
-
-	/**
-	 * @param passwordFile
-	 *            the full path name of a file containing a password to use with
-	 *            sudo (or empty for none)
-	 */
-	public abstract void setPasswordFile(String passwordFile);
-
-	/**
-	 * @return the full path name of a file containing a password to use with
-	 *         sudo (or empty for none)
-	 */
-	public abstract String getPasswordFile();
-
-	/**
-	 * @param operatingLimit
-	 *            the maximum number of runs in the
-	 *            {@linkplain Status#Operating operating} state at once
-	 */
-	public abstract void setOperatingLimit(int operatingLimit);
 }
